+++ .github/workflows/python-controller.yml
@@ -84,6 +84,15 @@ jobs:
         flags: python
         fail_ci_if_error: false
 
+    - name: Upload test results to Codecov
+      if: ${{ !cancelled() }}
+      uses: codecov/test-results-action@v1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        files: controller/results.xml
+        flags: python
+        fail_ci_if_error: false
+
 
   ci-status-python-controller:
     needs: [build-docker, run-checks]
