+++ car/conftest.py
@@ -1,0 +1,1 @@
+# pytest adds this directory to sys.path, making car-level helpers importable.
