+++ car/pytest.ini
@@ -1,1 +1,2 @@
 [pytest]
+pythonpath = test_helpers
