+++ car/test_apps/conftest.py
@@ -1,2 +1,1 @@
-# Shared fixtures and helpers for integration and E2E tests.
-# pytest adds this directory to sys.path, making helpers.py importable.
+# Shared fixtures for integration and E2E tests.
