+++ controller/pyproject.toml
@@ -22,7 +22,7 @@ legacy_tox_ini = """
 
     [testenv]
     deps = -r requirements/dev.txt
-    commands = pytest tests/ --cov=src --cov-report=xml:coverage.xml {posargs}
+    commands = pytest tests/ --cov=src --cov-report=xml:coverage.xml --junitxml=results.xml -o junit_family=legacy {posargs}
 """
 
 [tool.mypy]
